requirements.txt
branca==0.8.1
Flask==3.1.1
flask-bcrypt==1.0.1
flask-caching==2.3.1
flask-login==0.6.3
folium==0.20.0
geopandas==1.1.0
geopy==2.4.1
googlemaps==4.10.0
matplotlib==3.10.3
numpy==2.3.0
pandas==2.3.0
Pillow==11.2.1
pymongo==4.13.2
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
torch==2.7.1
transformers==4.52.4
webdriver-manager==4.0.2
shapely
fiona
rtree
openpyxl==3.1.2
gunicorn

